% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_channel.r
\name{Channel}
\alias{Channel}
\title{A Class which represents a channel object}
\usage{
Channel(data)
}
\description{
A Class which represents a channel object
}
\section{Slots}{

\describe{
\item{\code{id}}{the id of this channel}

\item{\code{type}}{the type of channel}

\item{\code{guild_id}}{the id of the guild}

\item{\code{position}}{sorting position of the channel}

\item{\code{permission_overwrites}}{explicit permission overwrites for members and roles}

\item{\code{name}}{the name of the channel (2-100 characters)}

\item{\code{topic}}{the channel topic (0-1024 characters)}

\item{\code{nsfw}}{whether the channel is nsfw}

\item{\code{last_message_id}}{the id of the last message sent in this channel (may not point to an existing or valid message)}

\item{\code{bitrate}}{the bitrate (in bits) of the voice channel}

\item{\code{user_limit}}{the user limit of the voice channel}

\item{\code{rate_limit_per_user}}{amount of seconds a user has to wait before sending another message}

\item{\code{recipients}}{the recipients of the DM}

\item{\code{icon}}{icon hash}

\item{\code{owner_id}}{id of the DM creator}

\item{\code{application_id}}{application id of the group DM creator if it is bot-created}

\item{\code{parent_id}}{id of the parent category for a channel}

\item{\code{last_pin_timestamp}}{when the last pinned message was pinned}
}}

\section{Disclaimer}{
 Pirate's channel object won't be separated to types!
see https://discordapp.com/developers/docs/resources/channel#channel-object-channel-types
to be able differentiate.
also, because of that, many fields are subject of being NULL accordingly.
}

