% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_client.r
\name{Client}
\alias{Client}
\title{A Class which represents the Client}
\usage{
Client(t)
}
\description{
A Class which represents the Client
}
\section{Slots}{

\describe{
\item{\code{token}}{The Client's token}

\item{\code{ws}}{The WebSocket connection of the client}

\item{\code{user}}{The client's user object}

\item{\code{private}}{set of fields you should not touch!}

\item{\code{guilds}}{The cached guilds of the Client}

\item{\code{direct_channels}}{The cached DMs of the Client}

\item{\code{users}}{The cached users of the Client}

\item{\code{presences}}{The cached presences of the Client}

\item{\code{emitter}}{The event emitter}
}}

\section{Warning}{

Do not mutate heartbeat value
}

