% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateway.r
\name{connect}
\alias{connect}
\title{Connect and authenticate to the websocket via a client}
\usage{
connect(client)
}
\arguments{
\item{client}{A Client class/object}
}
\description{
Connect and authenticate to the websocket via a client
}
\examples{
\dontrun{
client <- Client("<PRECIOUS TOKEN>")
connect(client)
}
}
