% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_emoji.r
\name{Emoji}
\alias{Emoji}
\title{A Class which represents a guild emoji object}
\usage{
Emoji(data, guild)
}
\description{
A Class which represents a guild emoji object
}
\section{Slots}{

\describe{
\item{\code{id}}{emoji id}

\item{\code{name}}{emoji name}

\item{\code{roles}}{roles this emoji is whitelisted to}

\item{\code{user}}{user that created this emoji}

\item{\code{require_colons}}{whether this emoji must be wrapped in colons}

\item{\code{managed}}{whether this emoji is managed}

\item{\code{animated}}{whether this emoji is animated}

\item{\code{guild_id}}{the guild_id this emoji belongs to}
}}

