% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_guildemojisupdate.r
\name{event.guild_emojis_update}
\alias{event.guild_emojis_update}
\title{Event, emitted whenever an Emoji is being updated}
\usage{
event.guild_emojis_update(data, client)
}
\arguments{
\item{data}{The event fields}

\item{client}{The client object}
}
\description{
Event, emitted whenever an Emoji is being updated
}
\section{Disclaimer}{

The way this event works is:
 This event calls sub events, by default the gateway doesn't have the option
 of recognizing what's an emoji remove and an emoji update for example.
 so I decided to handle that for you, adding 3 events in total of emojis update events
(These events are not documented in the official docs!)
'GUILD_EMOJIS_UPDATE'
'GUILD_EMOJIS_ADD'
'GUILD_EMOJIS_REMOVE'
}

