% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_emitter.r
\name{EventEmitter}
\alias{EventEmitter}
\title{A Class which emits event, heavily modelled after Node.js}
\usage{
EventEmitter(c)
}
\description{
A Class which emits event, heavily modelled after Node.js
}
\section{Slots}{

\describe{
\item{\code{client}}{The client to emmit the events to}

\item{\code{events}}{The event handlers}

\item{\code{emit}}{The function which is used to emit events}

\item{\code{on}}{The function which will add handlers}
}}

