% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_channelcreate.r
\name{events.channel_create}
\alias{events.channel_create}
\title{Event, emitted whenever a channel is being created}
\usage{
events.channel_create(data, client)
}
\arguments{
\item{data}{The event fields}

\item{client}{The client object}
}
\description{
Event, emitted whenever a channel is being created
}
\section{Disclaimer}{

Since Pirate does NOT handle channel types for you
you will have to be aware of it when involving complex actions
}

\examples{
\dontrun{
client$emitter$on("CHANNEL_CREATE", function(channel) {
 guild_channel_types = c(0, 2, 4, 5, 6)
 if (channel$type \%in\% guild_channel_types) {
   guild <- client$guilds$get(channel$guild_id)
   cat("A new channel called:", channel$name, "has been created in:", guild$id)
 }
})
}
}
