% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_channeldelete.r
\name{events.channel_delete}
\alias{events.channel_delete}
\title{Event, emitted whenever a channel is being deleted}
\usage{
events.channel_delete(data, client)
}
\arguments{
\item{data}{The event fields}

\item{client}{The client object}
}
\description{
Event, emitted whenever a channel is being deleted
}
\examples{
\dontrun{
client$emitter$on("CHANNEL_DELETE", function (channel) {
  cat(channel$name, "has been deleted, damnit!")
})
}

}
