% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_channelupdate.r
\name{events.channel_update}
\alias{events.channel_update}
\title{Event, emitted whenever a channel gets an update}
\usage{
events.channel_update(data, client)
}
\arguments{
\item{data}{The event fields}

\item{client}{The client object}
}
\description{
Event, emitted whenever a channel gets an update
}
\examples{
\dontrun{
client$emitter$on("CHANNEL_UPDATE", function(old_channel, new_channel) {
 # Assuming the name changed
 cat("New name:", new_channel$name, "Old name:", old_channel$name)
})
}

}
