% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_guildbanadd.r
\name{events.guild_ban_add}
\alias{events.guild_ban_add}
\title{Event, emitted whenever a user gets hammered by the power of thor.}
\usage{
events.guild_ban_add(data, client)
}
\description{
Event, emitted whenever a user gets hammered by the power of thor.
}
\section{Slots}{

\describe{
\item{\code{data}}{The event fields}

\item{\code{client}}{The client object}
}}

\examples{
\dontrun{
client$emitter$on("GUILD_BAN_ADD", function(guild, user) {
 cat("The user", user$username, "has been banned from", guild$name)
})
}
}
