% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_guildbanremove.r
\name{events.guild_ban_remove}
\alias{events.guild_ban_remove}
\title{Event, emitted whenever a user ban is being revoked}
\usage{
events.guild_ban_remove(data, client)
}
\arguments{
\item{data}{The event fields}

\item{client}{The client object}
}
\description{
Event, emitted whenever a user ban is being revoked
}
\section{Disclaimer}{

Be aware that whenever the guild won't be cached
the guild parameter will return as a guild id
guild id can be used to fetch the guild from the API
}

\examples{
\dontrun{
 client$emitter$on("GUILD_BAN_REMOVE", function(guild, user) {
 cat(user$name, "'s' ban has been revoked on", guild$name)
})
}
}
