% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_guildcreate.r
\name{events.guild_create}
\alias{events.guild_create}
\title{Event, emitted whenever the bot joins a guild}
\usage{
events.guild_create(data, client)
}
\arguments{
\item{data}{The new guild object}

\item{client}{The client object}
}
\description{
Event, emitted whenever the bot joins a guild
}
