% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_guildintegrationsupdate.r
\name{events.guild_integrations_update}
\alias{events.guild_integrations_update}
\title{Event, emitted whenever a guild's integrations is being updated}
\usage{
events.guild_integrations_update(data, client)
}
\arguments{
\item{data}{The event fields}

\item{client}{The client object}
}
\description{
Event, emitted whenever a guild's integrations is being updated
}
\section{Disclaimer}{

This event will return guild id instead of guild object if not cached.
this can be used in order to fetch the guild from the API
}

\examples{
\dontrun{
client$emitter$on("GUILD_INTEGRATIONS_UPDATE", function(guild) {
 cat(guild$name, "Just updated their integrations")
})
}
}
