% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_guildmemberadd.r
\name{events.guild_member_add}
\alias{events.guild_member_add}
\title{Event, emitted whenever a user joins a guild}
\usage{
events.guild_member_add(data, client)
}
\arguments{
\item{data}{The event fields}

\item{client}{The client fields}
}
\description{
Event, emitted whenever a user joins a guild
}
\section{Disclaimer}{

This event will return guild id instead of guild object if not cached.
this can be used in order to fetch the guild from the API
}

\examples{
\dontrun{
client$emitter$on("GUILD_MEMBER_ADD", function(guild, member) {
 cat(member$user$username, "just joined", guild$name)
})
}
}
