% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_guildmemberremove.r
\name{events.guild_member_remove}
\alias{events.guild_member_remove}
\title{Event, emits whenever a user leaves a guild}
\usage{
events.guild_member_remove(data, client)
}
\arguments{
\item{data}{The event fields}

\item{client}{The client object}
}
\description{
Event, emits whenever a user leaves a guild
}
\section{Disclaimer}{

This event will return guild id instead of guild object if not cached.
this can be used in order to fetch the guild from the API
}

\examples{
\dontrun{
client$emitter$on("GUILD_MEMBER_REMOVE", function(guild, member) {
 cat(member$user$username, "just left", guild$name)
})
}
}
