% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_guildmemberupdate.r
\name{events.guild_member_update}
\alias{events.guild_member_update}
\title{Event, emits whenever a member updates it's state}
\usage{
events.guild_member_update(data, client)
}
\arguments{
\item{data}{The event fields}

\item{client}{The client object}
}
\description{
Event, emits whenever a member updates it's state
}
\section{Disclaimer}{

This event will return guild id instead of guild object if not cached.
this can be used in order to fetch the guild from the API
AND old_member will return as NA if not cached
}

\section{Differences}{

This event will not return differences because it will be too expensive
of an operation.
}

\examples{
\dontrun{
client$emitter$on("GUILD_MEMBER_UPDATE", function(guild, old_member, new_member) {
 cat("Old nick", old_member$nick, "New:", new_member$nick)
})
}
}
