% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_guildmemberschunk.r
\name{events.guild_members_chunk}
\alias{events.guild_members_chunk}
\title{Event, emitted whenever one sends a respond for GUILD_MEMBERS_CHUNK}
\usage{
events.guild_members_chunk(data, client)
}
\arguments{
\item{data}{The event fields}

\item{client}{The client object}
}
\description{
Event, emitted whenever one sends a respond for GUILD_MEMBERS_CHUNK
}
\section{Disclaimer}{

This event will return guild id instead of guild object if not cached.
this can be used in order to fetch the guild from the API
If the guild is not cached therefore the chunk won't be cached to
as a result, the chunk will be returned as an array of GuildMember objects :)
so you will be able to cache the chunk when you fetch a new guild!
IF the guild is cached and it should be, the guild will be a guild object and
the members will be an array of IDs representing the new members
}

\examples{
\dontrun{
#TODO


}
}
