% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_guildrolecreate.r
\name{events.guild_role_create}
\alias{events.guild_role_create}
\title{Event, emitted whenever a role is being created}
\usage{
events.guild_role_create(data, client)
}
\arguments{
\item{data}{The event fields}

\item{client}{The client object
\dontrun{
client$emitter$on("GUILD_ROLE_CREATE", function(guild, role) {
 cat("A new role:", role$name, "has been created in:", guild$name, "!")
})
}}
}
\description{
Event, emitted whenever a role is being created
}
\section{Disclaimer}{

This event will return guild id instead of guild object if not cached.
this can be used in order to fetch the guild from the API (DISABLED FOR NOW)
}

