% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_guildupdate.r
\name{events.guild_update}
\alias{events.guild_update}
\title{Event, emitted whenever a guild updates it's contents}
\usage{
events.guild_update(data, client)
}
\arguments{
\item{data}{The updated guild object (partial)}

\item{client}{The client object}
}
\description{
Event, emitted whenever a guild updates it's contents
}
