% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_messagecreate.r
\name{events.message_create}
\alias{events.message_create}
\title{Event, emitted whenever one sends a message}
\usage{
events.message_create(data, client)
}
\arguments{
\item{data}{The event fields}

\item{client}{The client object}
}
\description{
Event, emitted whenever one sends a message
}
\examples{
\dontrun{
client$emitter$on("MESSAGE_CREATE", function(message) {
cat("Content:", message$content)
})
}
}
