% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_messagedelete.r
\name{events.message_delete}
\alias{events.message_delete}
\title{Event, emitted whenever a message is being deleted}
\usage{
events.message_delete(data, client)
}
\arguments{
\item{data}{The event fields}

\item{client}{The client object}
}
\description{
Event, emitted whenever a message is being deleted
}
\examples{
\dontrun{
client$emitter$on("MESSAGE_DELETE", function(channel, message) {
cat("Content of the deleted message:", message$content)
})
}
}
