% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_messagedeletebulk.r
\name{events.message_delete_bulk}
\alias{events.message_delete_bulk}
\title{Event, emitted whenever a message is being deleted}
\usage{
events.message_delete_bulk(data, client)
}
\arguments{
\item{data}{The event fields}

\item{client}{The client object}
}
\description{
Event, emitted whenever a message is being deleted
}
\examples{
\dontrun{
client$emitter$on("MESSAGE_DELETE_BULK", function(channel, messages) {
cat(length(messages), "have been deleted")
cat("Content of the deleted message:", message$content)
})
}
}
