% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_messagereactionadd.r
\name{events.message_reaction_add}
\alias{events.message_reaction_add}
\title{Event, emitted whenever a reaction is being attached to a message}
\usage{
events.message_reaction_add(data, client)
}
\arguments{
\item{data}{The event fields}

\item{client}{The client object}
}
\description{
Event, emitted whenever a reaction is being attached to a message
}
\examples{
\dontrun{
client$emitter$on("MESSAGE_REACTION_ADD", function(message, reaction) {
 cat(reaction$emoji$name, "has been added to", message$content)
})
}

}
