% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_messagereactionremove.r
\name{events.message_reaction_remove}
\alias{events.message_reaction_remove}
\title{Event, emitted whenever a reaction is being detached from a message}
\usage{
events.message_reaction_remove(data, client)
}
\arguments{
\item{data}{The event fields}

\item{client}{The client object}
}
\description{
Event, emitted whenever a reaction is being detached from a message
}
\examples{
\dontrun{
client$emitter$on("MESSAGE_REACTION_REMOVE", function(message, reaction) {
 cat(reaction$emoji$name, "has been removed from", message$content)
})
}

}
