% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_ready.r
\name{events.ready}
\alias{events.ready}
\title{Event, emitted whenever the client finished the identifying phase}
\usage{
events.ready(data, client)
}
\arguments{
\item{data}{Chunks of data relevant to the client}

\item{client}{The client object}
}
\description{
Event, emitted whenever the client finished the identifying phase
}
