% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_typingstart.r
\name{events.typing_start}
\alias{events.typing_start}
\title{Event, emitted whenever one starts to type}
\usage{
events.typing_start(data, client)
}
\arguments{
\item{data}{The event fields}

\item{client}{The client object}
}
\description{
Event, emitted whenever one starts to type
}
\examples{
\dontrun{
client$emitter$on("TYPING_START", channel, user) {
 cat("User:", user$username, "started smashing his keyboard in:", channel$name)
})
}
}
