% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_func.R
\name{group_by_func}
\alias{group_by_func}
\title{Summarise a Dataframe
This function takes a column from a dataframe, groups by it and summarises the result of the columns
The user can choose between "mean", "min", "max" and "sum" as summarisation functions}
\usage{
group_by_func(data, groupcol, func)
}
\arguments{
\item{data}{dataframe input}

\item{groupcol}{colname to group by}

\item{func}{summarisation function}
}
\value{
newdata asummarised dataframe
}
\description{
Summarise a Dataframe
This function takes a column from a dataframe, groups by it and summarises the result of the columns
The user can choose between "mean", "min", "max" and "sum" as summarisation functions
}
