% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_guild.r
\name{Guild}
\alias{Guild}
\title{A Class which represents a guild object}
\usage{
Guild(data)
}
\description{
A Class which represents a guild object
}
\section{Slots}{

\describe{
\item{\code{id}}{guild id}

\item{\code{name}}{guild name (2-100 characters)}

\item{\code{icon}}{icon hash}

\item{\code{splash}}{splash hash}

\item{\code{owner}}{whether or not the user is the owner of the guild}

\item{\code{owner_id}}{id of owner}

\item{\code{permissions}}{total permissions for the user in the guild (does not include channel overrides)}

\item{\code{region}}{voice region id for the guild}

\item{\code{afk_channel_id}}{id of afk channel}

\item{\code{afk_timeout}}{afk timeout in seconds}

\item{\code{embed_enabled}}{is this guild embeddable (e.g. widget)}

\item{\code{embed_channel_id}}{if not null, the channel id that the widget will generate an invite to}

\item{\code{verification_level}}{verification level required for the guild}

\item{\code{default_message_notifications}}{default message notifications level}

\item{\code{explicit_content_filter}}{explicit content filter level}

\item{\code{roles}}{roles in the guild}

\item{\code{emojis}}{custom guild emojis}

\item{\code{features}}{enabled guild features}

\item{\code{mfa_level}}{required MFA level for the guild}

\item{\code{application_id}}{application id of the guild creator if it is bot-created}

\item{\code{widget_enabled}}{whether or not the server widget is enabled}

\item{\code{widget_channel_id}}{the channel id for the server widget}

\item{\code{system_channel_id}}{the id of the channel to which system messages are sent}

\item{\code{joined_at}}{when this guild was joined at}

\item{\code{large}}{whether this is considered a large guild}

\item{\code{unavailable}}{is this guild unavailable}

\item{\code{member_count}}{total number of members in this guild}

\item{\code{voice_stated}}{(without the guild_id key)}

\item{\code{members}}{users in the guild}

\item{\code{channels}}{channels in the guild}

\item{\code{max_presences}}{the maximum amount of presences for the guild (the default value, currently 5000, is in effect when null is returned)}

\item{\code{max_members}}{the maximum amount of members for the guild}

\item{\code{vanity_url_code}}{the vanity url code for the guild}

\item{\code{description}}{the description for the guild}

\item{\code{banner}}{banner hash}

\item{\code{premium_tier}}{premium_tier (SEE: https://discordapp.com/developers/docs/resources/guild#guild-object-premium-tier)}

\item{\code{premium_subscription_count}}{the total number of users currently boosting this server}
}}

