% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_guildbanadd.r
\name{guildbanadd}
\alias{guildbanadd}
\title{Event, emitted whenever a user gets hammered by the power of thor.}
\usage{
guildbanadd(data, client)
}
\description{
Event, emitted whenever a user gets hammered by the power of thor.
}
\section{Slots}{

\describe{
\item{\code{data}}{The event fields}

\item{\code{client}}{The client object}
}}

\examples{
\dontrun{
client$emitter$on("GUILD_BAN_ADD", function(guild, user) {
 cat("The user", user$username, "has been banned from", guild$name)
})
}
}
