% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_guildcreate.r
\name{guildcreate}
\alias{guildcreate}
\title{Event, emitted whenever the bot joins a guild}
\usage{
guildcreate(data, client)
}
\arguments{
\item{data}{The new guild object}

\item{client}{The client object}
}
\description{
Event, emitted whenever the bot joins a guild
}
