% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_guildremove.r
\name{guildremove}
\alias{guildremove}
\title{whenever a guild becomes unavilable or the bot has been removed from there}
\usage{
guildremove(data, client)
}
\arguments{
\item{data}{The guild object (partial)}

\item{client}{The client object}
}
\description{
whenever a guild becomes unavilable or the bot has been removed from there
}
\examples{
\dontrun{
 client$emitter$on("GUILD_REMOVE", function(guild) {
   cat("The bot left ", guild$name)
 })
}
}
