% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_guildupdate.r
\name{guildupdate}
\alias{guildupdate}
\title{Event, emitted whenever a guild updates it's contents}
\usage{
guildupdate(data, client)
}
\arguments{
\item{data}{The updated guild object (partial)}

\item{client}{The client object}
}
\description{
Event, emitted whenever a guild updates it's contents
}
