% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_message.r
\name{Message}
\alias{Message}
\title{A Class which represents a message object}
\usage{
Message(data, client)
}
\description{
A Class which represents a message object
}
\section{Slots}{

\describe{
\item{\code{id}}{id of the message}

\item{\code{channel_id}}{id of the channel the message was sent in}

\item{\code{guild_id}}{id of the guild the message was sent in}

\item{\code{author}}{the author of this message (not guaranteed to be a valid user, see below)}

\item{\code{member}}{member properties for this message's author}

\item{\code{content}}{contents of the message}

\item{\code{timestamp}}{when this message was sent}

\item{\code{edited_timestamp}}{when this message was edited (or null if never)}

\item{\code{tts}}{whether this was a TTS message}

\item{\code{mention_everyone}}{whether this message mentions everyone}

\item{\code{mentions}}{users specifically mentioned in the message}

\item{\code{mention_roles}}{roles specifically mentioned in this message}

\item{\code{attachments}}{any attached files}

\item{\code{embeds}}{any embedded content}

\item{\code{reactions}}{reactions to the message}

\item{\code{nonce}}{used for validating a message was sent}

\item{\code{pinned}}{whether this message is pinned}

\item{\code{webhook_id}}{if the message is generated by a webhook, this is the webhook's id}

\item{\code{type}}{type of message}

\item{\code{activity}}{sent with Rich Presence-related chat embeds}

\item{\code{application}}{sent with Rich Presence-related chat embeds}
}}

