\name{pirate-package}
\alias{pirate-package}
\alias{pirate}
\docType{package}
\title{
\packageTitle{pirate}
}
\description{
\packageDescription{pirate}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pirate}
\packageIndices{pirate}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{pirate}

Maintainer: \packageMaintainer{pirate}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
