% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_presence.r
\name{Presence}
\alias{Presence}
\title{A Class which represents a presence object}
\usage{
Presence(data, g)
}
\description{
A Class which represents a presence object
}
\section{Slots}{

\describe{
\item{\code{user}}{the user presence is being updated for}

\item{\code{roles}}{roles this user is in}

\item{\code{game}}{null, or the user's current activity}

\item{\code{guild_id}}{id of the guild}

\item{\code{status}}{either "idle", "dnd", "online", or "offline"}

\item{\code{activities}}{user's current activities}

\item{\code{client_status}}{user's platform-dependent status}
}}

