% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_reaction.r
\name{Reaction}
\alias{Reaction}
\title{A Class which represents a reaction model}
\usage{
Reaction(data, client)
}
\description{
A Class which represents a reaction model
}
\section{Slots}{

\describe{
\item{\code{user_id}}{the id of the user}

\item{\code{channel_id}}{the id of the channel_id}

\item{\code{message_id}}{the id of the message}

\item{\code{guild_id}}{the id of the guild}

\item{\code{emoji}}{the emoji used to react}

\item{\code{custom}}{whether this is a custom emoji}
}}

