% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolvers.r
\name{resolve.snowflake}
\alias{resolve.snowflake}
\title{A function which resolves an ID to a timestamp in ms}
\usage{
resolve.snowflake(id)
}
\arguments{
\item{id}{The ID you want to resolve}
}
\description{
A function which resolves an ID to a timestamp in ms
}
