% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_role.r
\name{Role}
\alias{Role}
\title{A Class which represents a role object (exists in guilds only)}
\usage{
Role(data)
}
\description{
A Class which represents a role object (exists in guilds only)
}
\section{Slots}{

\describe{
\item{\code{id}}{role id}

\item{\code{name}}{role name}

\item{\code{color}}{integer representation of hexadecimal color code}

\item{\code{hoist}}{if this role is pinned in the user listing}

\item{\code{position}}{position of this role}

\item{\code{permissions}}{permissions bit set}

\item{\code{managed}}{whether this role is managed by an integration}

\item{\code{mentionable}}{whether this role is mentionable}
}}

