% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_updateclient.r
\name{updates.presence}
\alias{updates.presence}
\title{A Function which is being used to update the client's presence}
\usage{
updates.presence(client, presence)
}
\arguments{
\item{client}{The client object}

\item{presence}{The presence object (named list)}
}
\description{
A Function which is being used to update the client's presence
}
\section{Resource}{


See https://discordapp.com/developers/docs/topics/gateway#update-status
}

\section{Warning}{

Sending malformed payload will result in errors! don't panick :)
}

\examples{
\dontrun{update.presence(client, list(game=list(name="Hey!", type=0)))}
\dontrun{update.presence(client, list(game=list(name="Streaming!", type=1, url="https://twitch.tv/theonlyartz")))}
}
