% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_user.r
\docType{data}
\name{User}
\alias{User}
\title{A Class which represents a user object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
User
}
\description{
A Class which represents a user object
}
\section{id The ID of the user}{
NA
}

\section{username The username of the user}{
NA
}

\section{discriminator The discriminator of the user (#XXXX)}{
NA
}

\section{avatar The user's avatar hash}{
NA
}

\section{flags The flags on a user's account}{
NA
}

\section{premium_type The type of Nitro subscription on a user's account}{
NA
}

\keyword{datasets}
