% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.r
\docType{data}
\name{Client}
\alias{Client}
\title{A Class which represents the Client}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Client
}
\description{
A Class which represents the Client
}
\section{Slots}{

\describe{
\item{\code{token}}{The Client's token}

\item{\code{ws}}{The WebSocket connection of the client}

\item{\code{user}}{#TODO}

\item{\code{heartbeat}}{The heartbeat interval of the client}

\item{\code{lastSequence}}{The last sequence ID/Number from Discord}

\item{\code{guilds}}{The cached guilds of the Client}

\item{\code{users}}{The cached users of the Client}

\item{\code{presences}}{The cached presences of the Client}
}}

\section{Warning}{

Do not mutate heartbeat value
}

\keyword{datasets}
