% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_guildmemberupdate.r
\name{events.guild_member_update}
\alias{events.guild_member_update}
\title{Event, emits whenever a member updates it's state}
\usage{
events.guild_member_update(data, client)
}
\arguments{
\item{data}{The event fields}

\item{client}{The client object}
}
\description{
Event, emits whenever a member updates it's state
}
\section{Disclaimer}{

This event will return guild id instead of guild object if not cached.
this can be used in order to fetch the guild from the API
}

\examples{




}
