% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_guildmember.r
\docType{data}
\name{GuildMember}
\alias{GuildMember}
\title{A Class which represents a guild member}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GuildMember
}
\description{
A Class which represents a guild member
}
\section{Slots}{

\describe{
\item{\code{user}}{the user this guild member represents}

\item{\code{nick}}{this users guild nickname (if one is set)}

\item{\code{roles}}{array of role object ids}

\item{\code{joined_at}}{when the user joined the guild}

\item{\code{premium_since}}{when the user used their Nitro boost on the server}

\item{\code{deaf}}{whether the user is deafened in voice channels}

\item{\code{mute}}{whether the user is muted in voice channels}
}}

\keyword{datasets}
