% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_user.r
\docType{data}
\name{User}
\alias{User}
\title{A Class which represents a user object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
User
}
\description{
A Class which represents a user object
}
\section{Slots}{

\describe{
\item{\code{id}}{The ID of the user}

\item{\code{username}}{The username of the user}

\item{\code{discriminator}}{The discriminator of the user (#XXXX)}

\item{\code{avatar}}{The user's avatar hash}

\item{\code{flags}}{The flags on a user's account}

\item{\code{premium_type}}{The type of Nitro subscription on a user's account}
}}

\keyword{datasets}
